Remove commented out code; Add JWT auth to UsersController delete action

+++ server/controllers/UsersController.js
@@ -36,25 +36,20 @@ module.exports = (app) => {
 			.catch((error) => {
 				return res.status(400).json(error);
 			});
-		// if (!user.success) {
-		// 	return res.status(400).json(user);
-		// }
-
-		// User.query()
-		// 	.insertAndFetch(user.data)
-		// 	.then((data) => {
-		// 		if (data) {
-		// 			return res.json(data);
-		// 		} else {
-		// 			return res.status(400).json('User creation failed');
-		// 		}
-		// 	})
-		// 	.catch((error) => {
-		// 		return res.status(500).json(error);
-		// 	});
 	});
 
 	app.delete('/api/users/:id(\\d+)', (req, res) => {
+		var token = req.header.token,
+			id = req.params.id;
+
+		if (!AuthHelper.authenticateUserWithId(id, token)) {
+			return res.status(403).json({
+				success: false,
+				status: 403,
+				errors: ['Invalid authentication']
+			});
+		}
+
 		User.query()
 			.deleteById(req.params.id)
 			.then((data) => {
